Fix strtod ("NAN(I)") in Turkish locales (bug 19266).
The implementations of strtod and related functions use
locale-specific conversions to lower case when parsing the contents of
a string NAN(n-char-sequence_opt).  This has the consequence that
NAN(I) is not treated as being of that form (only the initial NAN part
is accepted).  The syntax of n-char-sequence directly maps to the
ASCII letters, digits and underscore as in identifiers, so it is
unambiguous that all ASCII letters must be accepted in all locales.

This patch, relative to a tree with
<https://sourceware.org/ml/libc-alpha/2015-11/msg00258.html> (pending
review) applied and depending on that patch, fixes this problem by
checking directly for ASCII letters.  This will have the side effect
of no longer accepting 'İ' (dotted 'I') inside NAN() in Turkish
locales, which seems appropriate (that letter wouldn't have been
interpreted as having any meaning in the NaN payload anyway, as not
acceptable to strtoull).

Tested for x86_64 and x86.

	[BZ #19266]
	* stdlib/strtod_l.c (____STRTOF_INTERNAL): Check directly for
	upper case and lower case letters inside NAN(), not using TOLOWER.
	* stdlib/tst-strtod-nan-locale-main.c: New file.
	* stdlib/tst-strtod-nan-locale.c: Likewise.
	* stdlib/Makefile (tests): Add tst-strtod-nan-locale.
	[$(run-built-tests) = yes] ($(objpfx)tst-strtod-nan-locale.out):
	Depend on $(gen-locales).
	($(objpfx)tst-strtod-nan-locale): Depend on $(libm).
	* wcsmbs/tst-wcstod-nan-locale.c: New file.
	* wcsmbs/Makefile (tests): Add tst-wcstod-nan-locale.
	[$(run-built-tests) = yes] ($(objpfx)tst-wcstod-nan-locale.out):
	Depend on $(gen-locales).
	($(objpfx)tst-wcstod-nan-locale): Depend on $(libm).