Fix ldbl-128ibm logl inaccuracy near 1 (bug 19351).
The ldbl-128ibm implementation of logl is inaccurate for arguments
near 1, because when deciding whether to bypass a series expansion for
log(1+z), where z = x-1, it compares the square of z rather than z
itself with an epsilon value.  This patch fixes that comparison, so
eliminating the test failures for inaccuracy of logl in such cases.

Tested for powerpc.

	[BZ #19351]
	* sysdeps/ieee754/ldbl-128ibm/e_logl.c (__ieee754_logl): When
	expanding log(1+z), compare z rather than its square with epsilon
	to determine when to avoid evaluating the expansion.