localedata: LC_IDENTIFICATION.category: set to ISO 30112 2014 standard
The ISO 30112 standard defines the valid values for the category
keyword as only a few options:
	posix:1993
	i18n:2004
	i18n:2012

The vast majority of locales had changed the "i18n" string to the
name of its own locale (e.g. "ak_GH:2013") as well as tweaking the
date (presumably thinking it should be the date of submission).

Convert all of them to "i18n:2012" for consistency.  A follow up
change will update localedef to actually check/validate the field.